+++ src/app/styles/checkbox-style/checkbox-style.component.spec.ts
@@ -1,0 +1,24 @@
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { IonicModule } from '@ionic/angular';
+
+import { CheckboxStyleComponent } from './checkbox-style.component';
+
+describe('CheckboxStyleComponent', () => {
+  let component: CheckboxStyleComponent;
+  let fixture: ComponentFixture<CheckboxStyleComponent>;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      declarations: [ CheckboxStyleComponent ],
+      imports: [IonicModule.forRoot()]
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(CheckboxStyleComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  }));
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});