+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - add [capacitor/app-launcher](https://capacitorjs.com/docs/apis/app-launcher)
  - add style `openAppButton` - open app based on the configuration, if the app is not installed it opens the installation link if there is one
  - use `initialView`, `headerToolbar_start`, `headerToolbar_center` and `headerToolbar_end` from the config in style `calendar`
+ - add `radioStyle` to be load in the basic style
 
 # 3.1.10
  - load `toggle switch` when it is enabled for a input from type `checkbox`